% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvec.R
\name{getvec}
\alias{getvec}
\title{read a physionet signal from file}
\usage{
getvec(filename)
}
\arguments{
\item{filename}{The path to the file to read, with or without extension}
}
\description{
read a physionet signal from file
}
\note{
The C function in wfdb must be called from the directory where the file resides. This wrapper make sure to move to the directory first, reads the signal and return to the original working directory
}

