% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getvec_}
\alias{getvec_}
\title{read the signals from a MIT database file}
\usage{
getvec_(filename)
}
\arguments{
\item{filename}{The path to the signal file, without extension}
}
\value{
a Matrix object with one signal per row
}
\description{
read the signals from a MIT database file
}
\note{
This function needs to be called from the directory where the file resides. Use getvec to go over this restriction
}

